requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.1.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
clr-loader==0.2.4
contourpy==1.0.6
cycler==0.11.0
dateparser==1.1.4
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
multidict==6.0.3
newnewtulipy==0.4.6.5
numpy==1.24.0
packaging==22.0
pandas==1.5.2
pandas-ta==0.3.14b0
pep517==0.13.0
Pillow==9.3.0
plotly==5.11.0
pycparser==2.21
pyparsing==3.0.9
python-binance==1.0.16
python-dateutil==2.8.2
python-telegram-bot==13.15
pythonnet==3.0.0.post1
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.1
six==1.16.0
stock-indicators==1.0.0
TA-Lib @ file:///C:/Users/vosni/projects/tradebot_v2/TA_Lib-0.4.24-cp310-cp310-win_amd64.whl
tenacity==8.1.0
tomli==2.0.1
tornado==6.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.6.0
urllib3==1.26.13
websockets==10.4
yarl==1.8.2
